requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pdfminer.six==20221105
python-docx==1.1.0
python-dotenv==1.0.1
